+fixed make tbz +showing max attr cache entries number in stats +cleanup
git-svn-id: https://svn.code.sf.net/p/remotefs/code/trunk@1008 7764f161-3e89-4957-9ae6-8e1f4830ae4a

+++ src/instance_client.h
@@ -44,7 +44,7 @@ struct rfs_instance
 #ifdef RFS_DEBUG
 		unsigned long cache_hits;
 		unsigned long cache_misses;
-		unsigned long max_number_of_entries;
+		unsigned max_number_of_entries;
 #endif
 	} attr_cache;
